backend/audos_console/requirements.txt
Flask==3.0.3
reportlab==4.2.2
Jinja2==3.1.4
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0